Fixed #10

+++ src/headers/render/X3D_texture.h
@@ -42,15 +42,15 @@ static inline int x3d_texture_h(const X3D_Texture* tex) {
     return tex->h;
 }
 
-static inline uint32 x3d_texture_pixel_index(const X3D_Texture* tex, uint16 u, uint16 v) {
-    return (uint32)v * tex->w + u;
+static inline uint32 x3d_texture_pixel_index(const X3D_Texture* tex, int u, int v) {
+    return v * tex->w + u;
 }
 
-static inline X3D_ColorIndex x3d_texture_get_texel(const X3D_Texture* tex, uint16 u, uint16 v) {  
+static inline X3D_ColorIndex x3d_texture_get_texel(const X3D_Texture* tex, int u, int v) {  
     return tex->texels[x3d_texture_pixel_index(tex, u, v)];
 }
 
-static inline void x3d_texture_set_texel(X3D_Texture* tex, uint16 u, uint16 v, X3D_ColorIndex c) {
+static inline void x3d_texture_set_texel(X3D_Texture* tex, int u, int v, X3D_ColorIndex c) {
   tex->texels[x3d_texture_pixel_index(tex, u, v)] = c;
 }
 
@@ -62,13 +62,13 @@ static inline void x3d_texture_init_empty(X3D_Texture* tex) {
 
 static inline _Bool x3d_texture_is_empty(const X3D_Texture* tex) {
     return tex->texels == NULL;
-}
+}   
 
 static inline uint32 x3d_texture_total_texels(const X3D_Texture* tex) {
     return (uint32)tex->w * tex->h;
 }
 
-static inline int16 x3d_texture_texel_is_valid(const X3D_Texture* tex, int16 u, int16 v) {
+static inline int16 x3d_texture_texel_is_valid(const X3D_Texture* tex, int u, int v) {
     return u >= 0 && v >= 0 && u < tex->w && v < tex->h;
 }
